Update Secrets Structure
